fixed critical bug breaking nuzlocke captures, idk how i didn't see it before???

+++ include/pokemon.h
@@ -247,7 +247,7 @@ extern const u8 gStatStageRatios[MAX_STAT_STAGE + 1][2];
 extern const u16 gLinkPlayerFacilityClasses[];
 extern const struct SpriteTemplate gBattlerSpriteTemplates[];
 extern const s8 gNatureStatTable[][5];
-extern const u32 gEvolutionLines[NUM_SPECIES][EVOS_PER_LINE];
+extern const u16 gEvolutionLines[NUM_SPECIES][EVOS_PER_LINE];
 extern const u16 *const gFormSpeciesIdTables[NUM_SPECIES];
 
 void ZeroBoxMonData(struct BoxPokemon *boxMon);
